Remove code for 'push' diagnostic tagging
