+++ src/EditorFeatures/CSharpTest/SuggestionTags/SuggestionTagProducerTests.cs
@@ -13,7 +13,6 @@ using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Editor.Shared.Extensions;
 using Microsoft.CodeAnalysis.Editor.UnitTests.Squiggles;
 using Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces;
-using Microsoft.CodeAnalysis.Options;
 using Microsoft.CodeAnalysis.Test.Utilities;
 using Microsoft.CodeAnalysis.Text;
 using Microsoft.VisualStudio.Text.Tagging;
@@ -26,8 +25,8 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.SuggestionTags
     [Trait(Traits.Feature, Traits.Features.SuggestionTags), Trait(Traits.Feature, Traits.Features.Tagging)]
     public class SuggestionTagProducerTests
     {
-        [WpfTheory, CombinatorialData]
-        public async Task SuggestionTagTest1(bool pull)
+        [WpfFact]
+        public async Task SuggestionTagTest1()
         {
             var (spans, selection) = await GetTagSpansAndSelectionAsync(
 @"class C {
@@ -35,16 +34,14 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.SuggestionTags
         var v = [|ne|]w X();
         v.Y = 1;
     }
-}", pull);
+}");
             Assert.Equal(1, spans.Length);
             Assert.Equal(selection, spans.Single().Span.Span.ToTextSpan());
         }
 
-        private static async Task<(ImmutableArray<ITagSpan<IErrorTag>> spans, TextSpan selection)> GetTagSpansAndSelectionAsync(
-            string content, bool pull)
+        private static async Task<(ImmutableArray<ITagSpan<IErrorTag>> spans, TextSpan selection)> GetTagSpansAndSelectionAsync(string content)
         {
             using var workspace = TestWorkspace.CreateCSharp(content);
-            workspace.GlobalOptions.SetGlobalOption(DiagnosticTaggingOptionsStorage.PullDiagnosticTagging, pull);
 
             var analyzerMap = new Dictionary<string, ImmutableArray<DiagnosticAnalyzer>>()
             {
