+++ src/EditorFeatures/Core/Diagnostics/AbstractDiagnosticsAdornmentTaggerProvider.cs
@@ -13,7 +13,7 @@ using Microsoft.VisualStudio.Text.Tagging;
 namespace Microsoft.CodeAnalysis.Diagnostics
 {
     internal abstract partial class AbstractDiagnosticsAdornmentTaggerProvider<TTag> :
-        AbstractPushOrPullDiagnosticsTaggerProvider<TTag>
+        AbstractDiagnosticsTaggerProvider<TTag>
         where TTag : class, ITag
     {
         protected AbstractDiagnosticsAdornmentTaggerProvider(
